more deliberate location for script
I was about to move this under /ansible-managed/ for consistency.  But the
script is run from a service, which may be subject to SELinux enforcement.
In this case, scripts under /ansible-managed/ will not necessarily have
a SELinux label that allows this.

/etc/ was _permitted_ for this, but not everyone is so happy about that
nowadays.  New daemons tend to go in /usr/libexec, so that seems like a
reasonable for now.

If we had several scripts, we could create a subdirectory.

It might make sense to use /usr/local/libexec/ansible-managed/,
particularly e.g. if a script was just called "hdparm"...
but this is ok for now.